styling datepicker
